requirements.txt
kubernetes==26.1.0
numpy==1.22
urllib3==1.25.4
PyYAML~=5.4.1